### 2017-06-03 #273 @rasbt
> @rasbt if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, this list of people have already agreed to review for JOSS and may be suitable for this submission.

Looking at the list, I think @krother @genomematt and @ctb are good candidates! Anyone of you interested and open to review the biopandas package?
