### 2017-06-05 #279 @whedon
[REVIEW]: BioPandas: Working with molecular structures in pandas DataFrames
**Submitting author:** @rasbt (<a href="http://orcid.org/0000-0001-6989-4493">Sebastian Raschka</a>)
**Repository:** <a href="https://github.com/rasbt/biopandas" target ="_blank">https://github.com/rasbt/biopandas</a>
**Version:** v0.2.2
