### 2017-06-05 52dc75c
travis for old software (#32)
