+++ .travis.yml
@@ -5,6 +5,7 @@ env:
   matrix:
     - PYTHON_VERSION="2.7" LATEST="true"
     - PYTHON_VERSION="3.5" LATEST="true"
+    - PYTHON_VERSION="3.6" NUMPY_VERSION="1.11.2" SCIPY_VERSION="0.18.1" PANDAS_VERSION="0.19.1"
     - PYTHON_VERSION="3.6" COVERAGE="true" LATEST="true"
 install: source ./ci/.travis_install.sh
 script: bash ./ci/.travis_test.sh
