### 2017-06-05 #279 @krother
The PDB parsing works out of the box and does not require further explanation. It is a killer app in itself. I found a (probably small) issue in the RMSD calculation feature.
At this moment, I am unable to establish the functional claims "reading and parsing millions of small molecule structures (from multi-MOL2 files)" and "filtering molecules by the presence of functional groups" based on the documentation.
