### 2017-06-05 #279 @krother
DOI references for the PDB could be:
Nature Structural Biology 10, 980 (2003); doi: 10.1038/nsb1203-980
and
The Protein Data Bank
Helen M. Berman
John Westbrook
Zukang Feng
Gary Gilliland
T. N. Bhat
Helge Weissig
Ilya N. Shindyalov
Philip E. Bourne
Nucleic Acids Res (2000) 28 (1): 235-242.
DOI: https://doi.org/10.1093/nar/28.1.235
