### 2017-06-05 #279 @krother
For pandas, a book written by Wes exists (http://shop.oreilly.com/product/0636920023784.do), but it has no DOI. I'd leave the pandas and MOL reference as they are.
