### 2017-06-05 #279 @rasbt
Thanks for the feedback!

> The PDB parsing works out of the box and does not require further explanation. It is a killer app in itself. I found a (probably small) issue in the RMSD calculation feature.

I saw a note about that issue in the biopandas repo and will take a look at it tonight and hopefully resolve it.

> At this moment, I am unable to establish the functional claims "reading and parsing millions of small molecule structures (from multi-MOL2 files)" and "filtering molecules by the presence of functional groups" based on the documentation.

When I wrote this I had another paper in mind that I was working on recently. I am just waiting for the co-author feedback and then the manuscript is going to be submitted to JCIM ~ next week. But I guess it will probably take a bit longer until that's published and citable. There, I used biopandas as an engine to perform filtering by functional groups (spatial distance between functional groups, charge range, and hybridization). I only have a brief section on that in the docs here:

http://rasbt.github.io/biopandas/tutorials/Working_with_MOL2_Structures_in_DataFrames/#parsing-multi-mol2-files

In this example, you can think of the `# do some analysis` placeholder as data frame operations like: does this molecule have a O.2 hybridized and S.o2 hybridized atom? If yes, are those within a distance of X A ... and so forth using and chaining in-built pandas/biopandas functions. I am happy to extend this section with a simple case if necessary. Would that be okay with you? I probably won't do the "million" there because processing millions of molecules could take a few hours (it's a short time for a research project, but maybe a bit to involved for a tutorial :P)
