### 2017-06-06 #279 @rasbt
> DOI references for the PDB could be: ...

Thanks, updated those references!
