### 2017-06-06 #279 @krother
argh, I didn't spot the menu bar on top on the documentation page. It's quite impressive and complete as it is, thanks! :+1:
