### 2017-06-06 330410e
rmsd with non-matching array indices
