### 2017-06-06 #279 @rasbt
> I found a (probably small) issue in the RMSD calculation feature.

This should be fixed in the master branch now (https://github.com/rasbt/biopandas/pull/34). For JOSS, I assume I need to make a new version release so that this change takes effect? It's not a problem, I think I should hold off a bit in case there's something else that needs to be addressed. (Once everything looks fine, I can make a new stable release, i.e., 2.1.0 -> 2.1.1 or 2.2.0). Is that okay?
