### 2017-06-06 #279 @krother
Checked the changed version by installing with pip directly from github. Great you put this in the documentation! And it works, too.
