### 2017-06-06 930bcbf
Added DOI to reference
