### 2017-06-06 #279 @krother
Created a PR where I indicated a possible location for the DOI in the PDB reference. No more issues from me. Congratulations to this contribution, it is time to get the word out!
