### 2017-06-06 #279 @rasbt
Thanks @krother . Just merged the DOI PR, and I am happy to hear that everything looks fine now :)
