### 2017-06-06 08e78ea
update PR template (#37)
