+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -12,7 +12,7 @@ Insert Description Here
 
 ### Related issues or pull requests
 
-<!-- Please provide a link to the respective issue on the [Issue Tracker](https://github.com/rasbt/mlxtend/issues) if one exists. E.g.,
+<!-- Please provide a link to the respective issue on the [Issue Tracker](https://github.com/rasbt/biopandas/issues) if one exists. E.g.,
 
 Fixes #<ISSUE_NUMBER> -->
 
@@ -22,13 +22,15 @@ Link related issues/pull requests here
 
 ### Pull Request requirements
 
-- [ ] Added appropriate unit test functions in the `./mlxtend/*/tests` directories
-- [ ] Ran `nosetests ./mlxtend -sv` and make sure that all unit tests pass
-- [ ] Checked the test coverage by running `nosetests ./mlxtend --with-coverage`
-- [ ] Checked for style issues by running `flake8 ./mlxtend`
+For new features or bug fixes, pleas consider the following to-do list:
+
+- [ ] Added appropriate unit test functions in the `./biopandas/*/tests` directories
+- [ ] Ran `nosetests ./biopandas -sv` and make sure that all unit tests pass
+- [ ] Checked the test coverage by running `nosetests ./biopandas --with-coverage`
+- [ ] Checked for style issues by running `flake8 ./biopandas`
 - [ ] Added a note about the modification or contribution to the `./docs/sources/`CHANGELOG.md` file
-- [ ] Modify documentation in the appropriate location under `mlxtend/docs/sources/` (optional)
-- [ ] Checked that the Travis-CI build passed at https://travis-ci.org/rasbt/mlxtend
+- [ ] Modified documentation in the appropriate location under `biopandas/docs/sources/` (optional)
+- [ ] Checked that the Travis-CI build passed at https://travis-ci.org/rasbt/biopandas
 
 
 
@@ -39,5 +41,5 @@ NOTE
 Due to the improved GitHub UI, the squashing of commits is no longer necessary.
 Please DO NOT SQUASH commits since they help with keeping track of the changes during the discussion).
 
-For more information and instructions, please see http://rasbt.github.io/mlxtend/contributing/
+For more information and instructions, please see http://rasbt.github.io/biopandas/contributing/
 -->
