### 2017-06-06 #279 @krother
OK, review done.
