+++ docs/sources/tutorials/Working_with_MOL2_Structures_in_DataFrames.md
@@ -1,31 +1,4 @@
 
-BioPandas
-
-Author: Sebastian Raschka <mail@sebastianraschka.com>  
-License: BSD 3 clause  
-Project Website: http://rasbt.github.io/biopandas/  
-Code Repository: https://github.com/rasbt/biopandas  
-
-
-```python
-%load_ext watermark
-%watermark -d -u -p pandas,biopandas
-```
-
-    last updated: 2017-04-02 
-    
-    pandas 0.19.2
-    biopandas 0.2.0.dev0
-
-
-
-```python
-from biopandas.mol2 import PandasMol2
-import pandas as pd
-pd.set_option('display.width', 600)
-pd.set_option('display.max_columns', 8)
-```
-
 # Working with MOL2 Structures in DataFrames
 
 The Tripos MOL2 format is a common format for working with small molecules. In this tutorial, we will go over some examples that illustrate how we can use Biopandas' MOL2 DataFrames to analyze molecules conveniently.
@@ -41,12 +14,16 @@ from biopandas.mol2 import PandasMol2
 pmol = PandasMol2().read_mol2('./data/1b5e_1.mol2')
 ```
 
+[File link: [1b5e_1.mol2](https://raw.githubusercontent.com/rasbt/biopandas/master/docs/sources/tutorials/data/1b5e_1.mol2)]
+
 The `read_mol2` method can also load structures from `.mol2.gz` files, but if you have a multi-mol2 file, keep in mind that it will only fetch the first molecule in this file. In the section "[Parsing Multi-MOL2 files](#parsing-multi-mol2-files)," we will see how we can parse files that contain multiple structures.
 
 
 ```python
 pmol = PandasMol2().read_mol2('./data/40_mol2_files.mol2.gz')
 ```
+
+[File link: [40_mol2_files.mol2.gz](https://github.com/rasbt/biopandas/blob/master/docs/sources/tutorials/data/40_mol2_files.mol2.gz?raw=true)]
 
 After the file was succesfully loaded, we have access to the following basic `PandasMol2` attributes:
 
@@ -358,6 +335,8 @@ pmol.df.tail(10)
 
 
 
+[File link: [1b5e_1.mol2](https://raw.githubusercontent.com/rasbt/biopandas/master/docs/sources/tutorials/data/1b5e_1.mol2)]
+
 For example, we can select all hydrogen atoms by filtering on the atom type column:
 
 
@@ -636,6 +615,8 @@ from biopandas.mol2 import PandasMol2
 pmol = PandasMol2().read_mol2('./data/1b5e_1.mol2')
 ```
 
+[File link: [1b5e_1.mol2](https://raw.githubusercontent.com/rasbt/biopandas/master/docs/sources/tutorials/data/1b5e_1.mol2)]
+
 
 ```python
 %matplotlib inline
@@ -655,7 +636,7 @@ plt.show()
 ```
 
 
-![png](Working_with_MOL2_Structures_in_DataFrames_files/Working_with_MOL2_Structures_in_DataFrames_30_0.png)
+![png](Working_with_MOL2_Structures_in_DataFrames_files/Working_with_MOL2_Structures_in_DataFrames_34_0.png)
 
 
 If this is too fine-grained for our needs, we could summarize the different atom types by atomic elements:
@@ -671,7 +652,7 @@ plt.show()
 ```
 
 
-![png](Working_with_MOL2_Structures_in_DataFrames_files/Working_with_MOL2_Structures_in_DataFrames_32_0.png)
+![png](Working_with_MOL2_Structures_in_DataFrames_files/Working_with_MOL2_Structures_in_DataFrames_36_0.png)
 
 
[…]
