### 2017-06-07 #279 @arfon
@rasbt - Could you move the references you currently have in the `paper.md` file into a `paper.bib` file and cite them directly please? (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))

When you've done that could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
