+++ paper.bib
@@ -1,0 +1,46 @@
+@inproceedings{mckinney2010data,
+author = {McKinney, Wes and Others},
+booktitle = {Proceedings of the 9th Python in Science Conference},
+mendeley-groups = {Scientific Computing},
+organization = {van der Voort S, Millman J},
+pages = {51--56},
+title = {{Data structures for statistical computing in python}},
+volume = {445},
+year = {2010}
+}
+@article{berman2003announcing,
+author = {Berman, Helen and Henrick, Kim and Nakamura, Haruki},
+journal = {Nature Structural {\&} Molecular Biology},
+mendeley-groups = {Papers/ScreenLamp Paper/software},
+number = {12},
+pages = {980},
+publisher = {Nature Publishing Group},
+title = {{Announcing the worldwide protein data bank}},
+volume = {10},
+year = {2003}
+}
+@article{Berman2000,
+abstract = {The Protein Data Bank (PDB; http://www.rcsb.org/pdb/ ) is the single worldwide archive of structural data of biological macromolecules. This paper describes the goals of the PDB, the systems in place for data deposition and access, how to obtain further information, and near-term plans for the future development of the resource.},
+author = {Berman, H. M.},
+doi = {10.1093/nar/28.1.235},
+file = {:Users/sebastian/Documents/Library/Mendeley/Berman et al/2000/Berman et al.{\_}2000{\_}The protein data bank.pdf:pdf},
+isbn = {1072-8368},
+issn = {13624962},
+journal = {Nucleic Acids Research},
+mendeley-groups = {Papers/SiteInterlock Paper,Papers/ScreenLamp Paper/software},
+month = {jan},
+number = {1},
+pages = {235--242},
+pmid = {10592235},
+publisher = {Oxford University Press},
+title = {{The Protein Data Bank}},
+url = {https://academic.oup.com/nar/article-lookup/doi/10.1093/nar/28.1.235},
+volume = {28},
+year = {2000}
+}
+@article{tripos2007tripos,
+  title={Tripos Mol2 File Format},
+  author={Tripos, L},
+  journal={St. Louis, MO: Tripos},
+  year={2007}
+}
