+++ biopandas/__init__.py
@@ -4,5 +4,5 @@
 # Project Website: http://rasbt.github.io/biopandas/
 # Code Repository: https://github.com/rasbt/biopandas
 
-__version__ = '0.2.2dev'
+__version__ = '0.2.2'
 __author__ = "Sebastian Raschka <mail@sebastianraschka.com>"
