### 2017-06-07 #279 @rasbt
@arfon  no problem.

I moved the references to a separate bibtex file, paper.bib.

> When you've done that could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive?

Here's the link for the 0.2.1 version that was reviewed:

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.574879.svg)](https://doi.org/10.5281/zenodo.574879)

This existed prior to the submission though and thus, it doesn't include the paper.md and paper.bib files.

 I made a new biopandas version (v0.2.1 -> v0.2.2) for zenodo incorporating the paper files and also including a small fix that came up during the review process. The zenodo link and doi:

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.804030.svg)](https://doi.org/10.5281/zenodo.804030)
